refactor: split phase visuals into per-component files
Each phase visual now lives in src/ui/tui/components/visualizer/, including
MatrixRain. PhaseVisuals.tsx is the orchestrator (phase dispatch + Winamp
wrapper + EQ helper); panel.tsx holds the shared Panel, VisualProps, and
MATRIX_FADE color.

Also: fix Visualizer breaking layout on short terminals (measureElement +
greedy chrome drop), stop slicing CrateStack package names, replace the
TrendLine dashboards visual with a 2×2 DashboardGrid of mini-charts.
